+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 CHANGELOG
 =========
 
+* fix a permission check that prevented the user from modifying the course
+  overview page
 * blocks for the overview page are not associated with a particular course
 * don't show video container if no video was embedded
 * fix how the URL of the test modification page is built